Veritas/projects/Veritas-frontend/requirements.txt
fastapi==0.115.6
python-multipart==0.0.20
imagehash==4.3.1
Pillow==11.1.0
numpy==2.2.2
scipy==1.15.1
requests==2.32.3
cloudinary==1.42.1
python-dotenv==1.0.1
